word_solver: Store the template and spare letters differently
Instead of storing a reference to the template it now owns a grid::Word.
This makes it easier to store the word solver in a struct. The spare
letters are now passed in directly as a Vec<char> instead of a string
because it will just be converted to that anyway.