Split the grid into two different representation formats
One format is the result of parsing the text representation. It doesn’t
split the data into words and doesn’t duplicate any of the letters. In
order solve the grid, there is now a WordGrid representation which
contains the list of words. The letters are just stored as Option<char>
and then there is a separate Vec of the movable letters without their
position in the grid.