grid: Always mark the first letter as yellow
If a puzzle has repeats of a letter that appears in the solution then it
will now always just mark the first one as yellow even if that means
that an intersecting word could also make it mark one of the repeats as
yellow as well. Previously the code would try to make it prefer
highlighting only the intersecting letter in that case but it didn’t
always work and it’s probably quite a hard problem to get right. It
seems like the original Waffle has the new behaviour so we might as well
do that.